fix(diracx-cli): tests client token authentication
